Improve logic.

+++ app/services/save_booking_details.rb
@@ -35,13 +35,11 @@ class SaveBookingDetails
         lodging = Lodging.friendly.find(reservations_attribute[:lodging_slug]) rescue nil
         reservation.lodging = lodging
 
-        if lodging.present? & reservation.save(validate: false) && booking.step_passed?(:booked) && !reservation.canceled?
+        if lodging.present? & reservation.save(validate: false) && booking.step_passed?(:booked) && !reservation.canceled? && !booking.rebooking_approved?
           lodging.availabilities.check_out_only!(reservation.check_in)
           lodging.availabilities.where(available_on: (reservation.check_in+1.day..reservation.check_out-1.day).map(&:to_s)).destroy_all
           lodging.availabilities.where(available_on: reservation.check_out, check_out_only: true).delete_all
         end
-
-        add_availabilities(reservation.check_in, reservation.check_out, lodging) if params[:booking][:rebooking_approved]
 
         next unless reservations_attribute[:review_attributes].present?
         review = reservation.review || reservation.build_review
@@ -164,12 +162,5 @@ class SaveBookingDetails
     def created_by value
       return value if Booking.created_bies.keys.include?(value)
       'nice2stay'
-    end
-
-    def add_availabilities check_in, check_out, lodging
-      dates = (check_in..check_out).map(&:to_s)
-      dates.each do |date|
-        a = Availability.find_or_create_by(available_on: date, lodging_id: lodging.id)
-      end
     end
 end
